Introduce INode3D interface
