+++ src/Leviathan/CMakeLists.txt
@@ -1,39 +1,53 @@
-include_directories("${IRRLICHT_INCLUDE_DIRS}" "${YAML_INCLUDE_DIRS}")
-
-add_library(leviathan SHARED
-  characters/hero.cpp
-  characters/heroes.cpp
-  core/Configuration.cpp
-  core/GameStateManager.cpp
-  core/Logger.cpp
-  core/Randomizer.cpp
-  core/TimeControl.cpp
-  core/Timer.cpp
-  gui/Menu.cpp
-  gui/MenuControl.cpp
-  gui/MousePointerControl.cpp
-  input/ActionMapping.cpp
-  input/Actions.cpp
-  input/EventReceiver.cpp
-  input/EventToActionConverter.cpp
-  input/GUIEventActions.cpp
-  input/Input.cpp
-  input/KeyboardEventActions.cpp
-  input/MouseEventActions.cpp
-  video/Camera.cpp
-  video/MeshHelper.cpp
-  video/Textures.cpp
-  world/Ground.cpp
-  LeviathanDevice.cpp
+cmake_minimum_required(VERSION 3.9.2)
+project(leviathan DESCRIPTION "The Leviathan game engine." VERSION 0.5.1 LANGUAGES CXX)
+include_directories("${IRRLICHT_INCLUDE_DIRS}" "${YAML_INCLUDE_DIRS}" "${PUBLIC_INTERFACES}")
+add_library(${PROJECT_NAME} SHARED)
+target_sources(${PROJECT_NAME}
+  PRIVATE
+    characters/hero.cpp
+    characters/heroes.cpp
+    core/Configuration.cpp
+    core/GameStateManager.cpp
+    core/Logger.cpp
+    core/Randomizer.cpp
+    core/TimeControl.cpp
+    core/Timer.cpp
+    gui/Menu.cpp
+    gui/MenuControl.cpp
+    gui/MousePointerControl.cpp
+    input/ActionMapping.cpp
+    input/Actions.cpp
+    input/EventReceiver.cpp
+    input/EventToActionConverter.cpp
+    input/GUIEventActions.cpp
+    input/Input.cpp
+    input/KeyboardEventActions.cpp
+    input/MouseEventActions.cpp
+    video/Camera.cpp
+    video/MeshHelper.cpp
+    video/Textures.cpp
+    world/Ground.cpp
+    LeviathanDevice.cpp
+    ${PUBLIC_HEADER_FILES}
 )
-
-target_link_libraries(leviathan "${IRRLICHT_LIBRARY}" "${YAML_LIBRARY}")
-set_target_properties(leviathan PROPERTIES
-  LIBRARY_OUTPUT_NAME leviathan
-  PREFIX ""
+target_include_directories(${PROJECT_NAME}
+  PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
+    ${PUBLIC_INTERFACES}
+  PUBLIC
+    $<BUILD_INTERFACE:${PUBLIC_INTERFACES}>
 )
-set(LEVIATHAN_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR} PARENT_SCOPE)
-cotire(leviathan)
+target_link_libraries(${PROJECT_NAME}
+  "${IRRLICHT_LIBRARY}"
+  "${YAML_LIBRARY}"
+)
+set_target_properties(${PROJECT_NAME}
+  PROPERTIES
+    LIBRARY_OUTPUT_NAME ${PROJECT_NAME}
+    PREFIX ""
+)
+set(LEVIATHAN_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR} ${PUBLIC_INTERFACES} PARENT_SCOPE)
+cotire(${PROJECT_NAME})
 
-install(TARGETS leviathan RUNTIME DESTINATION .)
+install(TARGETS ${PROJECT_NAME} RUNTIME DESTINATION .)
 install(FILES "${IRRLICHT_BINARY_LIBRARY}" DESTINATION .)
